Move `pkg/util/iptables` to `k8s.io/kube-proxy/util/iptables`
The CRI-O container runtime is on its path to stop vendoring the main
Kubernetes repository, while only the `pkg/util/iptables` package is the
last major dependency. I propose to move the utility package to the
kube-proxy staging repo, because for k/k the proxy is the main consumer
of it.
