+++ pkg/kubelet/kubelet_network_linux.go
@@ -26,8 +26,8 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
 	"k8s.io/klog/v2"
+	utiliptables "k8s.io/kube-proxy/util/iptables"
 	"k8s.io/kubernetes/pkg/features"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
 	utilexec "k8s.io/utils/exec"
 )
