+++ staging/src/k8s.io/kube-proxy/go.mod
@@ -5,8 +5,12 @@ module k8s.io/kube-proxy
 go 1.20
 
 require (
+	github.com/lithammer/dedent v1.1.0
+	golang.org/x/sys v0.8.0
 	k8s.io/apimachinery v0.0.0
 	k8s.io/component-base v0.0.0
+	k8s.io/klog/v2 v2.100.1
+	k8s.io/utils v0.0.0-20230406110748-d93618cff8a2
 )
 
 require (
@@ -24,8 +28,6 @@ require (
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 	gopkg.in/inf.v0 v0.9.1 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
-	k8s.io/klog/v2 v2.100.1 // indirect
-	k8s.io/utils v0.0.0-20230406110748-d93618cff8a2 // indirect
 	sigs.k8s.io/json v0.0.0-20221116044647-bc3834ca7abd // indirect
 	sigs.k8s.io/structured-merge-diff/v4 v4.2.3 // indirect
 	sigs.k8s.io/yaml v1.3.0 // indirect
